Split netcdf load/save attribute testing into separate sourcefile.
